Privacy Trim: centimeter storage, a 200m cap, and wiring it into ingest
users.privacy_trim_m stored whole meters, so an imperial account's feet
input drifted by a foot on save-and-reload (two compounding roundings).
Renamed the column to privacy_trim_cm and round to the nearest
centimeter instead of the nearest meter — exact for every whole-foot
value, since 1 ft = 30.48 cm exactly. Also lowered the sanity ceiling
from 5000m to 200m.

Separately, persistAndEnqueue was hardcoding a 200m trim for every
upload regardless of the account's own setting — Settings could be
changed and would echo back correctly, but never actually affected
ingest. requireAuth's authInfo now carries privacyTrimCm alongside
userID/timezone, resolved once per request, so a plain upload, a .zip
archive, a Takeout import, and a GPS Logger sync batch all trim by
whatever the account's Settings page currently holds. Confirmed live:
two accounts (trim=0 and trim=200m) uploading the same ~1000m track
landed at 998.87m and 598.87m respectively.

+++ services/server/internal/ingest/ingest.go
@@ -35,12 +35,12 @@ const simplifyToleranceDeg = 0.00003
 // Job is what the `ingest` job's payload carries — everything the worker needs that isn't
 // already in the raw payload itself.
 type Job struct {
-	UserID        string `json:"user_id"`
-	Source        string `json:"source"`
-	SourceDetail  string `json:"source_detail"`
-	ExternalID    string `json:"external_id"`
-	RawPayloadKey string `json:"raw_payload_key"`
-	PrivacyTrimM  int    `json:"privacy_trim_m"`
+	UserID        string  `json:"user_id"`
+	Source        string  `json:"source"`
+	SourceDetail  string  `json:"source_detail"`
+	ExternalID    string  `json:"external_id"`
+	RawPayloadKey string  `json:"raw_payload_key"`
+	PrivacyTrimM  float64 `json:"privacy_trim_m"`
 	// ActivityType overrides whatever the parser itself reports, when set. Empty means "use
 	// the parsed value" (the zero value already does the right thing for every existing
 	// caller). This exists for the Google Takeout import path
